Add thundering herd protection
See https://en.wikipedia.org/wiki/Thundering_herd_problem for details.
In essence if multiple checks are configured with an overlapping interval
(e.g. all with 5min or 5min & 10min), especially to the same system,
the service will get hammered/DoSed by the middleware. To prevent that,
before the job is started for the first time a small semi-random delay
is introduced. This will spread the load at least over multiple seconds.