Add support for HTTP negative contents match to http-middleware
While this is possible with a clever regex in CHECK_MATCH_CONTENT, it is error-prone.
In addition, not all "grep" engines support extended syntax, which is especially
true in Busybox-based multicalls. In addition, separating both options allows
for cleaner logging and flexibkle usage with both positive & negative matching
in a single ping instance.